--- crack-core/src/commands/music/play_utils/mod.rs
@@ -1,5 +1,0 @@
-pub(crate) mod query;
-pub(crate) mod queue;
-
-pub(crate) use query::*;
-pub(crate) use queue::*;
