+++ crack-core/src/music/mod.rs
@@ -1,3 +1,7 @@
 pub mod context;
+pub(crate) mod query;
+pub(crate) mod queue;
 
 pub use context::QueryContext;
+pub(crate) use query::*;
+pub(crate) use queue::*;
