+++ crack-core/src/music/query.rs
