+++ crack-core/src/music/queue.rs
